requirements-develop.txt
-r requirements.txt

pytest==3.3.*               # Testing
pytest-cov==2.5.*           # Coverage Integration
pytest-django==3.1.*        # Django Integration

flake8==3.5.*               # Linting
